move REST setup to setup.go
